styled MGByUser under user's profile and added a delete confirmation

+++ src/seller/ManageProductsByUser.js
@@ -4,7 +4,7 @@ import { isAuth, getCookie } from "../auth/helpers";
 import { Link } from "react-router-dom";
 import { getProducts, deleteProduct } from "./apiAdmin";
 
-const ManageProducts = ({ userId }) => {
+const ManageProductsByUser = ({ userId }) => {
   const [products, setProducts] = useState([]);
 
   // const { user, token } = isAuth();
@@ -21,14 +21,19 @@ const ManageProducts = ({ userId }) => {
     });
   };
 
-  const destroy = (productId) => {
-    deleteProduct(productId, user._id, token).then((data) => {
-      if (data.error) {
-        console.log(data.error);
-      } else {
-        loadProducts();
-      }
-    });
+  const destroy = async (productName, productId) => {
+    let answer = await window.confirm(
+      `Are you sure you want to delete the product "${productName}"?`
+    );
+    if (answer) {
+      deleteProduct(productId, user._id, token).then((data) => {
+        if (data.error) {
+          console.log(data.error);
+        } else {
+          loadProducts();
+        }
+      });
+    }
   };
 
   useEffect(() => {
@@ -36,34 +41,44 @@ const ManageProducts = ({ userId }) => {
   }, []);
 
   return (
-    <div className="col-12">
+    <>
       <h2 className="text-center">My products</h2>
+      <br />
       {products
         .filter((p) => p.soldBy === userId)
         .map((p, i) => (
-          <>
-            <ul className="list-group">
-              <li
-                key={i}
-                className="list-group-item d-flex justify-content-between align-items-center"
-              >
-                <strong>{p.name}</strong>
+          <div key={i}>
+            <div style={{ height: "40px" }} className="row bg-white mb-1 px-3">
+              <div style={{ color: "dark" }} className="col-10 mx-auto my-auto">
+                {p.name}
+              </div>
+
+              <div className="col-1  my-auto">
                 <Link to={`/admin/product/update/${p._id}`}>
-                  <span className="badge badge-warning badge-pill">Update</span>
+                  <span
+                    style={{ color: "orange" }}
+                    data-toggle="tooltip"
+                    data-placement="left"
+                    title="Edit product"
+                    className="fas fa-edit"
+                  ></span>
                 </Link>
-                <span
-                  onClick={() => destroy(p._id)}
-                  className="badge badge-danger badge-pill"
-                >
-                  Delete
-                </span>
-              </li>
-            </ul>
-          </>
+              </div>
+
+              <div
+                style={{ color: "red" }}
+                onClick={() => destroy(p.name, p._id)}
+                data-toggle="tooltip"
+                data-placement="left"
+                title="Delete product"
+                className="col-1 fas fa-trash-alt text-center my-auto"
+              ></div>
+            </div>
+          </div>
         ))}
       <br />
-    </div>
+    </>
   );
 };
 
-export default ManageProducts;
+export default ManageProductsByUser;
